types change to size_t

+++ ft_bzero.c
@@ -6,11 +6,13 @@
 /*   By: akostian <akostian@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/08 11:23:59 by akostian          #+#    #+#             */
-/*   Updated: 2024/04/08 14:05:02 by akostian         ###   ########.fr       */
+/*   Updated: 2024/04/11 03:24:12 by akostian         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-void	ft_bzero(void *s, unsigned long int n)
+#include "libft.h"
+
+void	ft_bzero(void *s, size_t n)
 {
 	char	*s_cpy;
