+++ ft_memcmp.c
@@ -6,11 +6,13 @@
 /*   By: akostian <akostian@student.s19.be>         +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/04/09 11:15:23 by akostian          #+#    #+#             */
-/*   Updated: 2024/04/09 11:23:16 by akostian         ###   ########.fr       */
+/*   Updated: 2024/04/11 03:25:45 by akostian         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
-int	ft_memcmp(const void *s1, const void *s2, unsigned long int n)
+#include "libft.h"
+
+int	ft_memcmp(const void *s1, const void *s2, size_t n)
 {
 	char	*s1_cpy;
 	char	*s2_cpy;
